[feat]: [CDS-78377] move duplicate code to common module
