+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/environment/helper/EnvironmentMapper.java
@@ -20,8 +20,6 @@ import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.annotations.dev.ProductModule;
 import io.harness.cdng.envGroup.beans.EnvironmentGroupEntity;
-import io.harness.cdng.environment.yaml.EnvironmentPlanCreatorConfig;
-import io.harness.data.structure.CollectionUtils;
 import io.harness.ng.core.environment.beans.Environment;
 import io.harness.ng.core.environment.beans.EnvironmentType;
 import io.harness.ng.core.environment.yaml.NGEnvironmentConfig;
@@ -51,52 +49,6 @@ import org.apache.commons.lang3.StringUtils;
 public class EnvironmentMapper {
   private static final List<ServiceOverridesType> reverseOverridePriority =
       List.of(ENV_GLOBAL_OVERRIDE, ENV_SERVICE_OVERRIDE, INFRA_GLOBAL_OVERRIDE, INFRA_SERVICE_OVERRIDE);
-  public EnvironmentStepParameters toEnvironmentStepParameters(
-      EnvironmentPlanCreatorConfig environmentPlanCreatorConfig) {
-    Map<String, Object> serviceOverrides = new HashMap<>();
-    if (areSvcOverrideVariablesPresent(environmentPlanCreatorConfig)) {
-      serviceOverrides = NGVariablesUtils.getMapOfVariables(
-          environmentPlanCreatorConfig.getServiceOverrideConfig().getServiceOverrideInfoConfig().getVariables());
-    }
-
-    return EnvironmentStepParameters.builder()
-        .environmentRef(environmentPlanCreatorConfig.getEnvironmentRef())
-        .name(environmentPlanCreatorConfig.getName())
-        .identifier(environmentPlanCreatorConfig.getIdentifier())
-        .description(environmentPlanCreatorConfig.getDescription())
-        .tags(environmentPlanCreatorConfig.getTags())
-        .type(environmentPlanCreatorConfig.getType())
-        .serviceOverrides(serviceOverrides)
-        .variables(NGVariablesUtils.getMapOfVariables(environmentPlanCreatorConfig.getVariables()))
-        .build();
-  }
-
-  private boolean areSvcOverrideVariablesPresent(EnvironmentPlanCreatorConfig environmentPlanCreatorConfig) {
-    return environmentPlanCreatorConfig != null && environmentPlanCreatorConfig.getServiceOverrideConfig() != null
-        && environmentPlanCreatorConfig.getServiceOverrideConfig().getServiceOverrideInfoConfig() != null
-        && environmentPlanCreatorConfig.getServiceOverrideConfig().getServiceOverrideInfoConfig().getVariables()
-        != null;
-  }
-
-  public EnvironmentOutcome toEnvironmentOutcome(EnvironmentStepParameters stepParameters) {
-    overrideServiceVariables(stepParameters.getVariables(), stepParameters.getServiceOverrides());
-    return EnvironmentOutcome.builder()
-        .identifier(stepParameters.getIdentifier())
-        .name(stepParameters.getName() != null ? stepParameters.getName() : "")
-        .description(stepParameters.getDescription() != null ? stepParameters.getDescription() : "")
-        .tags(CollectionUtils.emptyIfNull(stepParameters.getTags()))
-        .type(stepParameters.getType())
-        .v1Type(stepParameters.getType() == EnvironmentType.Production ? "PROD" : "NON_PROD")
-        .environmentRef(stepParameters.getEnvironmentRef().getValue())
-        .variables(stepParameters.getVariables())
-        .build();
-  }
-
-  private void overrideServiceVariables(Map<String, Object> variables, Map<String, Object> serviceOverrides) {
-    if (variables != null && serviceOverrides != null) {
-      variables.putAll(serviceOverrides);
-    }
-  }
 
   private Map<String, Object> overrideVariables(List<NGVariable> base, List<NGVariable> override) {
     if (isEmpty(base)) {
