+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/service/steps/ServiceStepV3.java
@@ -6,12 +6,12 @@
  */
 
 package io.harness.cdng.service.steps;
+
 import static io.harness.cdng.gitops.constants.GitopsConstants.GITOPS_ENV_OUTCOME;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.ENV_GROUP_REF;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.ENV_REF;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.ENV_VARIABLES_PATTERN_REGEX;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.OVERRIDES_COMMAND_UNIT;
-import static io.harness.cdng.service.steps.constants.ServiceStepConstants.OVERRIDE_IN_REVERSE_PRIORITY;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.SERVICE_CONFIGURATION_NOT_FOUND;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.SERVICE_STEP_COMMAND_UNIT;
 import static io.harness.cdng.service.steps.constants.ServiceStepConstants.SERVICE_VARIABLES_PATTERN_REGEX;
@@ -19,7 +19,6 @@ import static io.harness.data.structure.CollectionUtils.emptyIfNull;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 import static io.harness.eraro.ErrorCode.FREEZE_EXCEPTION;
-import static io.harness.ng.core.environment.mappers.EnvironmentMapper.toNGEnvironmentConfig;
 import static io.harness.ng.core.serviceoverridev2.beans.ServiceOverridesType.ENV_GLOBAL_OVERRIDE;
 import static io.harness.ng.core.serviceoverridev2.beans.ServiceOverridesType.ENV_SERVICE_OVERRIDE;
 
@@ -31,7 +30,6 @@ import io.harness.annotations.dev.HarnessModuleComponent;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.annotations.dev.ProductModule;
-import io.harness.beans.common.VariablesSweepingOutput;
 import io.harness.cdng.artifact.outcome.ArtifactsOutcome;
 import io.harness.cdng.configfile.ConfigFilesOutcome;
 import io.harness.cdng.envGroup.beans.EnvironmentGroupEntity;
@@ -40,7 +38,6 @@ import io.harness.cdng.environment.helper.EnvironmentInfraFilterHelper;
 import io.harness.cdng.environment.helper.EnvironmentMapper;
 import io.harness.cdng.environment.helper.EnvironmentPlanCreatorHelper;
 import io.harness.cdng.environment.helper.EnvironmentStepsUtils;
-import io.harness.cdng.expressions.CDExpressionResolver;
 import io.harness.cdng.freeze.FreezeOutcome;
 import io.harness.cdng.gitops.steps.EnvClusterRefs;
 import io.harness.cdng.gitops.steps.GitOpsEnvOutCome;
@@ -54,6 +51,7 @@ import io.harness.cdng.service.beans.NativeHelmServiceSpec;
 import io.harness.cdng.service.steps.constants.ServiceStepV3Constants;
 import io.harness.cdng.service.steps.helpers.ServiceOverrideUtilityFacade;
 import io.harness.cdng.service.steps.helpers.ServiceStepsHelper;
+import io.harness.cdng.service.steps.helpers.ServiceStepsHelper.ServicePartResponse;
 import io.harness.cdng.service.steps.helpers.beans.ServiceStepV3Parameters;
 import io.harness.cdng.stepsdependency.constants.OutcomeExpressionConstants;
 import io.harness.cdng.visitor.YamlTypes;
@@ -76,9 +74,7 @@ import io.harness.logging.UnitStatus;
 import io.harness.logstreaming.NGLogCallback;
 import io.harness.ng.core.common.beans.NGTag;
 import io.harness.ng.core.environment.beans.Environment;
-import io.harness.ng.core.environment.beans.NGEnvironmentGlobalOverride;
 import io.harness.ng.core.environment.services.EnvironmentService;
-import io.harness.ng.core.environment.services.impl.EnvironmentEntityYamlSchemaHelper;
 import io.harness.ng.core.environment.yaml.NGEnvironmentConfig;
 import io.harness.ng.core.environment.yaml.NGEnvironmentInfoConfig;
 import io.harness.ng.core.service.entity.ServiceEntity;
@@ -93,7 +89,6 @@ import io.harness.ng.core.serviceoverride.yaml.NGServiceOverrideConfig;
 import io.harness.ng.core.serviceoverride.yaml.NGServiceOverrideInfoConfig;
 import io.harness.ng.core.serviceoverridev2.beans.NGServiceOverrideConfigV2;
 import io.harness.ng.core.serviceoverridev2.beans.ServiceOverridesSpec;
-import io.harness.ng.core.serviceoverridev2.beans.ServiceOverridesSpec.ServiceOverridesSpecBuilder;
 import io.harness.ng.core.serviceoverridev2.beans.ServiceOverridesType;
 import io.harness.ng.core.utils.ServiceOverrideV2ValidationHelper;
 import io.harness.pms.contracts.ambiance.Ambiance;
@@ -115,7 +110,6 @@ import io.harness.pms.sdk.core.steps.executables.ChildrenExecutable;
 import io.harness.pms.sdk.core.steps.io.StepInputPackage;
 import io.harness.pms.sdk.core.steps.io.StepResponse;
 import io.harness.pms.yaml.ParameterField;
-import io.harness.pms.yaml.YAMLFieldNameConstants;
 import io.harness.pms.yaml.YamlUtils;
 import io.harness.rbac.CDNGRbacUtility;
 import io.harness.steps.OutputExpressionConstants;
@@ -140,7 +134,6 @@ import com.google.inject.name.Named;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumMap;
 import java.util.HashMap;
@@ -151,9 +144,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
@@ -166,7 +156,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
   @Inject private ServiceStepsHelper serviceStepsHelper;
   @Inject private ExecutionSweepingOutputService sweepingOutputService;
   @Inject private EnvironmentService environmentService;
-  @Inject private CDExpressionResolver expressionResolver;
   @Inject private ServiceOverrideService serviceOverrideService;
   @Inject private ServiceStepOverrideHelper serviceStepOverrideHelper;
   @Inject private FreezeEvaluateService freezeEvaluateService;
@@ -180,7 +169,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
   @Inject @Named("PRIVILEGED") private AccessControlClient accessControlClient;
   @Inject private ServiceCustomSweepingOutputHelper serviceCustomSweepingOutputHelper;
   @Inject private ServiceEntityYamlSchemaHelper serviceEntityYamlSchemaHelper;
-  @Inject private EnvironmentEntityYamlSchemaHelper environmentEntityYamlSchemaHelper;
 
   @Inject private ServiceOverrideUtilityFacade serviceOverrideUtilityFacade;
   @Inject private ServiceOverrideV2ValidationHelper overrideV2ValidationHelper;
@@ -203,20 +191,20 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
       final NGLogCallback overrideLogCallback =
           serviceStepsHelper.getServiceLogCallback(ambiance, true, OVERRIDES_COMMAND_UNIT);
 
-      saveExecutionLog(serviceStepLogCallback, "Starting service step...");
+      serviceStepsHelper.saveExecutionLog(serviceStepLogCallback, "Starting service step...");
 
       Map<FreezeEntityType, List<String>> entityMap = new HashMap<>();
 
       final ServicePartResponse servicePartResponse = executeServicePart(ambiance, stepParameters, entityMap);
 
-      saveExecutionLog(serviceStepLogCallback,
+      serviceStepsHelper.saveExecutionLog(serviceStepLogCallback,
           "Service Name: " + servicePartResponse.getNgServiceConfig().getNgServiceV2InfoConfig().getName()
               + " , Identifier: "
               + servicePartResponse.getNgServiceConfig().getNgServiceV2InfoConfig().getIdentifier());
 
       performFiltering(stepParameters, ambiance,
           EnvironmentInfraFilterHelper.getNGTags(
-              servicePartResponse.ngServiceConfig.getNgServiceV2InfoConfig().getTags()));
+              servicePartResponse.getNgServiceConfig().getNgServiceV2InfoConfig().getTags()));
       // Support GitOps Flow
       // If environment group is only set for GitOps or if GitOps flow and deploying to multi-environments
       if (ParameterField.isNotNull(stepParameters.getGitOpsMultiSvcEnvEnabled())
@@ -361,16 +349,17 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
       NGEnvironmentConfig ngEnvironmentConfig;
       // handle old environments
       if (isEmpty(environment.getYaml())) {
-        setYamlInEnvironment(environment);
+        serviceStepsHelper.setYamlInEnvironment(environment);
       }
       try {
         if (isNotEmpty(parameters.getEnvToEnvInputs())) {
-          ngEnvironmentConfig = mergeEnvironmentInputs(accountId, environment.getIdentifier(), environment.getYaml(),
-              parameters.getEnvToEnvInputs().get(
-                  getEnvRefOrId(environment.fetchRef(), parameters.getEnvGroupRef(), environment.getIdentifier())));
-        } else {
           ngEnvironmentConfig =
-              mergeEnvironmentInputs(accountId, environment.getIdentifier(), environment.getYaml(), null);
+              serviceStepsHelper.mergeEnvironmentInputs(accountId, environment.getIdentifier(), environment.getYaml(),
+                  parameters.getEnvToEnvInputs().get(
+                      getEnvRefOrId(environment.fetchRef(), parameters.getEnvGroupRef(), environment.getIdentifier())));
+        } else {
+          ngEnvironmentConfig = serviceStepsHelper.mergeEnvironmentInputs(
+              accountId, environment.getIdentifier(), environment.getYaml(), null);
         }
       } catch (IOException ex) {
         throw new InvalidRequestException(format("Unable to read yaml for environment [Name: %s, Identifier: %s]",
@@ -407,13 +396,14 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
 
     serviceStepsHelper.checkForAccessOrThrow(ambiance, secretNGVariables);
 
-    resolve(ambiance, envToEnvVariables, envToSvcVariables);
+    serviceStepsHelper.resolve(ambiance, envToEnvVariables, envToSvcVariables);
 
     GitOpsEnvOutCome gitOpsEnvOutCome = new GitOpsEnvOutCome(envToEnvVariables, envToSvcVariables);
 
     sweepingOutputService.consume(ambiance, GITOPS_ENV_OUTCOME, gitOpsEnvOutCome, StepCategory.STAGE.name());
 
-    processServiceVariables(ambiance, servicePartResponse, serviceStepLogCallback, null, false, new HashMap<>());
+    serviceStepsHelper.processServiceVariables(
+        ambiance, servicePartResponse, serviceStepLogCallback, null, false, new HashMap<>());
 
     serviceStepOverrideHelper.prepareAndSaveFinalManifestMetadataToSweepingOutput(
         servicePartResponse.getNgServiceConfig(), null, null, ambiance,
@@ -467,11 +457,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
     return environments;
   }
 
-  private void setYamlInEnvironment(Environment environment) {
-    NGEnvironmentConfig ngEnvironmentConfig = toNGEnvironmentConfig(environment);
-    environment.setYaml(io.harness.ng.core.environment.mappers.EnvironmentMapper.toYaml(ngEnvironmentConfig));
-  }
-
   private void executeEnvironmentPart(Ambiance ambiance, ServiceStepV3Parameters parameters,
       ServicePartResponse servicePartResponse, NGLogCallback serviceStepLogCallback,
       Map<FreezeEntityType, List<String>> entityMap, NGLogCallback overrideLogCallback) throws IOException {
@@ -486,7 +471,7 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
     }
 
     if (envRef.isExpression()) {
-      resolve(ambiance, envRef);
+      serviceStepsHelper.resolve(ambiance, envRef);
     }
 
     log.info("Starting execution for Environment Step [{}]", envRef.getValue());
@@ -508,7 +493,7 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
 
       if (isOverridesV2enabled) {
         if (!ParameterField.isNull(parameters.getInfraId()) && parameters.getInfraId().isExpression()) {
-          resolve(ambiance, parameters.getInfraId());
+          serviceStepsHelper.resolve(ambiance, parameters.getInfraId());
         }
         mergedOverrideV2Configs = serviceOverrideUtilityFacade.getMergedServiceOverrideConfigs(
             accountId, orgIdentifier, projectIdentifier, parameters, environment.get(), overrideLogCallback);
@@ -517,11 +502,11 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
       NGEnvironmentConfig ngEnvironmentConfig;
       // handle old environments
       if (isEmpty(environment.get().getYaml())) {
-        setYamlInEnvironment(environment.get());
+        serviceStepsHelper.setYamlInEnvironment(environment.get());
       }
 
       try {
-        ngEnvironmentConfig = mergeEnvironmentInputs(
+        ngEnvironmentConfig = serviceStepsHelper.mergeEnvironmentInputs(
             accountId, environment.get().getIdentifier(), environment.get().getYaml(), envInputs);
       } catch (IOException ex) {
         throw new InvalidRequestException(
@@ -537,20 +522,21 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
             mergeSvcOverrideInputs(ngServiceOverridesEntity.get().getYaml(), parameters.getServiceOverrideInputs());
       }
 
-      resolve(ambiance, ngEnvironmentConfig, ngServiceOverrides);
+      serviceStepsHelper.resolve(ambiance, ngEnvironmentConfig, ngServiceOverrides);
 
       List<NGVariable> secretNGVariables = new ArrayList<>();
       if (isOverridesV2enabled) {
         if (ngEnvironmentConfig != null && ngEnvironmentConfig.getNgEnvironmentInfoConfig() != null
             && !mergedOverrideV2Configs.containsKey(ENV_GLOBAL_OVERRIDE)) {
-          mergedOverrideV2Configs.put(ENV_GLOBAL_OVERRIDE, toOverrideConfigV2(ngEnvironmentConfig, accountId));
+          mergedOverrideV2Configs.put(
+              ENV_GLOBAL_OVERRIDE, serviceStepsHelper.toOverrideConfigV2(ngEnvironmentConfig, accountId));
         }
         if (ngServiceOverrides != null && ngServiceOverrides.getServiceOverrideInfoConfig() != null
             && !mergedOverrideV2Configs.containsKey(ENV_SERVICE_OVERRIDE)) {
           mergedOverrideV2Configs.put(
               ENV_SERVICE_OVERRIDE, toOverrideConfigV2(ngServiceOverrides, accountId, ngEnvironmentConfig));
         }
-        secretNGVariables = getSecretVariablesFromOverridesV2(mergedOverrideV2Configs);
+        secretNGVariables = serviceStepsHelper.getSecretVariablesFromOverridesV2(mergedOverrideV2Configs);
 
       } else {
         if (ngEnvironmentConfig != null && ngEnvironmentConfig.getNgEnvironmentInfoConfig() != null
@@ -592,8 +578,8 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
       sweepingOutputService.consume(
           ambiance, OutputExpressionConstants.ENVIRONMENT, environmentOutcome, StepCategory.STAGE.name());
 
-      processServiceVariables(ambiance, servicePartResponse, serviceStepLogCallback, environmentOutcome,
-          isOverridesV2enabled, mergedOverrideV2Configs);
+      serviceStepsHelper.processServiceVariables(ambiance, servicePartResponse, serviceStepLogCallback,
+          environmentOutcome, isOverridesV2enabled, mergedOverrideV2Configs);
 
       if (isOverridesV2enabled) {
         NGServiceV2InfoConfig ngServiceV2InfoConfig =
@@ -635,24 +621,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
     }
   }
 
-  @NonNull
-  private List<NGVariable> getSecretVariablesFromOverridesV2(
-      Map<ServiceOverridesType, NGServiceOverrideConfigV2> mergedOverrideV2Configs) {
-    return mergedOverrideV2Configs.values()
-        .stream()
-        .map(NGServiceOverrideConfigV2::getSpec)
-        .map(spec -> spec.getVariables())
-        .filter(variables -> isNotEmpty(variables))
-        .flatMap(Collection::stream)
-        .filter(SecretNGVariable.class ::isInstance)
-        .collect(Collectors.toList());
-  }
-
-  private void resolve(Ambiance ambiance, Object... objects) {
-    final List<Object> toResolve = new ArrayList<>(Arrays.asList(objects));
-    expressionResolver.updateExpressions(ambiance, toResolve);
-  }
-
   private NGServiceOverrideConfig mergeSvcOverrideInputs(
       String originalOverridesYaml, ParameterField<Map<String, Object>> serviceOverrideInputs) {
     NGServiceOverrideConfig serviceOverrideConfig = NGServiceOverrideConfig.builder().build();
@@ -666,36 +634,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
       serviceOverrideConfig = ServiceOverridesMapper.toNGServiceOverrideConfig(mergedYaml);
     }
     return serviceOverrideConfig;
-  }
-
-  private void processServiceVariables(Ambiance ambiance, ServicePartResponse servicePartResponse,
-      NGLogCallback serviceStepLogCallback, EnvironmentOutcome environmentOutcome, boolean isOverridesV2Enabled,
-      Map<ServiceOverridesType, NGServiceOverrideConfigV2> overridesV2Configs) {
-    VariablesSweepingOutput variablesSweepingOutput;
-    if (isOverridesV2Enabled) {
-      variablesSweepingOutput =
-          getVariablesSweepingOutputFromOverridesV2(servicePartResponse.getNgServiceConfig().getNgServiceV2InfoConfig(),
-              serviceStepLogCallback, overridesV2Configs);
-    } else if (environmentOutcome != null) {
-      variablesSweepingOutput =
-          getVariablesSweepingOutput(servicePartResponse.getNgServiceConfig().getNgServiceV2InfoConfig(),
-              serviceStepLogCallback, environmentOutcome);
-    } else {
-      variablesSweepingOutput = getVariablesSweepingOutputForGitOps(
-          servicePartResponse.getNgServiceConfig().getNgServiceV2InfoConfig(), serviceStepLogCallback);
-    }
-
-    sweepingOutputService.consume(ambiance, YAMLFieldNameConstants.VARIABLES, variablesSweepingOutput, null);
-
-    Object outputObj = variablesSweepingOutput.get("output");
-    if (!(outputObj instanceof VariablesSweepingOutput)) {
-      outputObj = new VariablesSweepingOutput();
-    }
-
-    sweepingOutputService.consume(ambiance, YAMLFieldNameConstants.SERVICE_VARIABLES,
-        (VariablesSweepingOutput) outputObj, StepCategory.STAGE.name());
-
-    saveExecutionLog(serviceStepLogCallback, "Processed service variables");
   }
 
   @Override
@@ -733,8 +671,8 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
     final NGLogCallback logCallback =
         serviceStepsHelper.getServiceLogCallback(ambiance, false, SERVICE_STEP_COMMAND_UNIT);
     if (StatusUtils.brokeStatuses().contains(stepResponse.getStatus())) {
-      saveExecutionLog(logCallback, LogHelper.color("Failed to complete service step", LogColor.Red), LogLevel.INFO,
-          CommandExecutionStatus.FAILURE);
+      serviceStepsHelper.saveExecutionLog(logCallback, LogHelper.color("Failed to complete service step", LogColor.Red),
+          LogLevel.INFO, CommandExecutionStatus.FAILURE);
       serviceStepUnitProgress = UnitProgress.newBuilder()
                                     .setStatus(UnitStatus.FAILURE)
                                     .setUnitName(SERVICE_STEP_COMMAND_UNIT)
@@ -742,7 +680,8 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
                                     .setEndTime(System.currentTimeMillis())
                                     .build();
     } else {
-      saveExecutionLog(logCallback, "Completed service step", LogLevel.INFO, CommandExecutionStatus.SUCCESS);
+      serviceStepsHelper.saveExecutionLog(
+          logCallback, "Completed service step", LogLevel.INFO, CommandExecutionStatus.SUCCESS);
       serviceStepUnitProgress = UnitProgress.newBuilder()
                                     .setStatus(UnitStatus.SUCCESS)
                                     .setUnitName(SERVICE_STEP_COMMAND_UNIT)
@@ -918,118 +857,6 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
         originalServiceYaml, YamlPipelineUtils.writeYamlString(serviceInputsYaml), true, true);
   }
 
-  private NGEnvironmentConfig mergeEnvironmentInputs(String accountId, String identifier, String yaml,
-      ParameterField<Map<String, Object>> environmentInputs) throws IOException {
-    if (ParameterField.isNull(environmentInputs) || isEmpty(environmentInputs.getValue())) {
-      return getNgEnvironmentConfig(accountId, identifier, yaml);
-    }
-    Map<String, Object> environmentInputYaml = new HashMap<>();
-    environmentInputYaml.put(YamlTypes.ENVIRONMENT_YAML, environmentInputs);
-    String resolvedYaml = MergeHelper.mergeRuntimeInputValuesAndCheckForRuntimeInOriginalYaml(
-        yaml, YamlPipelineUtils.writeYamlString(environmentInputYaml), true, true);
-
-    return getNgEnvironmentConfig(accountId, identifier, resolvedYaml);
-  }
-
-  private NGEnvironmentConfig getNgEnvironmentConfig(String accountId, String identifier, String yaml)
-      throws IOException {
-    try {
-      return YamlUtils.read(yaml, NGEnvironmentConfig.class);
-    } catch (Exception ex) {
-      environmentEntityYamlSchemaHelper.validateSchema(accountId, yaml);
-      log.error(String.format(
-          "Environment schema validation succeeded but failed to convert environment yaml to environment config [%s]",
-          identifier));
-      throw ex;
-    }
-  }
-
-  private VariablesSweepingOutput getVariablesSweepingOutput(
-      NGServiceV2InfoConfig serviceV2InfoConfig, NGLogCallback logCallback, EnvironmentOutcome environmentOutcome) {
-    // env v2 incorporating env variables into service variables
-    final Map<String, Object> envVariables = new HashMap<>();
-    if (isNotEmpty(environmentOutcome.getVariables())) {
-      envVariables.putAll(environmentOutcome.getVariables());
-    }
-    Map<String, Object> variables = getFinalVariablesMap(serviceV2InfoConfig, envVariables, logCallback);
-    VariablesSweepingOutput variablesOutcome = new VariablesSweepingOutput();
-    variablesOutcome.putAll(variables);
-    return variablesOutcome;
-  }
-
-  private VariablesSweepingOutput getVariablesSweepingOutputFromOverridesV2(NGServiceV2InfoConfig serviceV2InfoConfig,
-      NGLogCallback logCallback, Map<ServiceOverridesType, NGServiceOverrideConfigV2> overridesV2Configs) {
-    Map<String, Object> finalOverridesVariables = new HashMap<>();
-    final Map<String, Object> overridesVariables = getAllOverridesVariables(overridesV2Configs, logCallback);
-    if (isNotEmpty(overridesVariables)) {
-      finalOverridesVariables.putAll(overridesVariables);
-    }
-    Map<String, Object> variables = getFinalVariablesMap(serviceV2InfoConfig, finalOverridesVariables, logCallback);
-    VariablesSweepingOutput variablesOutcome = new VariablesSweepingOutput();
-    variablesOutcome.putAll(variables);
-    return variablesOutcome;
-  }
-
-  private Map<String, Object> getAllOverridesVariables(
-      Map<ServiceOverridesType, NGServiceOverrideConfigV2> overridesV2Configs, NGLogCallback logCallback) {
-    // Priority Order : INFRA_SERVICE > INFRA_GLOBAL > ENV_SERVICE > ENV_GLOBAL
-    Map<String, Object> finalOverridesVars = new HashMap<>();
-    for (ServiceOverridesType overridesType : OVERRIDE_IN_REVERSE_PRIORITY) {
-      if (overridesV2Configs.containsKey(overridesType)
-          && isNotEmpty(overridesV2Configs.get(overridesType).getSpec().getVariables())) {
-        finalOverridesVars.putAll(
-            NGVariablesUtils.getMapOfVariables(overridesV2Configs.get(overridesType).getSpec().getVariables()));
-        saveExecutionLog(logCallback, "Collecting variables from override of type " + overridesType.toString());
-      }
-    }
-    return finalOverridesVars;
-  }
-
-  private VariablesSweepingOutput getVariablesSweepingOutputForGitOps(
-      NGServiceV2InfoConfig serviceV2InfoConfig, NGLogCallback logCallback) {
-    Map<String, Object> variables = getFinalVariablesMap(serviceV2InfoConfig, Map.of(), logCallback);
-    VariablesSweepingOutput variablesOutcome = new VariablesSweepingOutput();
-    variablesOutcome.putAll(variables);
-    return variablesOutcome;
-  }
-
-  private Map<String, Object> getFinalVariablesMap(NGServiceV2InfoConfig serviceV2InfoConfig,
-      Map<String, Object> envOrOverrideVariables, NGLogCallback logCallback) {
-    List<NGVariable> variableList = serviceV2InfoConfig.getServiceDefinition().getServiceSpec().getVariables();
-    Map<String, Object> variables = new HashMap<>();
-    Map<String, Object> outputVariables = new VariablesSweepingOutput();
-    if (isNotEmpty(variableList)) {
-      Map<String, Object> originalVariables = NGVariablesUtils.getMapOfVariables(variableList);
-      variables.putAll(originalVariables);
-      outputVariables.putAll(originalVariables);
-    }
-    addEnvVariables(outputVariables, envOrOverrideVariables, logCallback);
-    variables.put("output", outputVariables);
-    return variables;
-  }
-
-  private void addEnvVariables(
-      Map<String, Object> variables, Map<String, Object> envVariables, NGLogCallback logCallback) {
-    if (isEmpty(envVariables)) {
-      return;
-    }
-
-    saveExecutionLog(logCallback, "Applying environment variables and service overrides");
-    variables.putAll(envVariables);
-  }
-
-  private void saveExecutionLog(NGLogCallback logCallback, String line) {
-    if (logCallback != null) {
-      logCallback.saveExecutionLog(line);
-    }
-  }
-
-  private void saveExecutionLog(NGLogCallback logCallback, String line, LogLevel info, CommandExecutionStatus success) {
-    if (logCallback != null) {
-      logCallback.saveExecutionLog(line, info, success);
-    }
-  }
-
   @VisibleForTesting
   protected ChildrenExecutableResponse executeFreezePart(
       Ambiance ambiance, Map<FreezeEntityType, List<String>> entityMap) {
@@ -1123,7 +950,7 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
         IdentifierRefHelper.getRefFromIdentifierOrRef(accountId, ngEnvironmentInfoConfig.getOrgIdentifier(),
             ngEnvironmentInfoConfig.getProjectIdentifier(), ngEnvironmentInfoConfig.getIdentifier());
     return NGServiceOverrideConfigV2.builder()
-        .identifier(generateEnvServiceOverrideV2Identifier(
+        .identifier(serviceStepsHelper.generateEnvServiceOverrideV2Identifier(
             serviceOverrideInfoConfig.getEnvironmentRef(), serviceOverrideInfoConfig.getServiceRef()))
         .environmentRef(envRef)
         .type(ENV_SERVICE_OVERRIDE)
@@ -1135,43 +962,5 @@ public class ServiceStepV3 implements ChildrenExecutable<ServiceStepV3Parameters
                   .applicationSettings(serviceOverrideInfoConfig.getApplicationSettings())
                   .build())
         .build();
-  }
-
-  private NGServiceOverrideConfigV2 toOverrideConfigV2(NGEnvironmentConfig envConfig, String accountId) {
-    NGEnvironmentInfoConfig ngEnvironmentInfoConfig = envConfig.getNgEnvironmentInfoConfig();
-    NGEnvironmentGlobalOverride ngEnvironmentGlobalOverride = ngEnvironmentInfoConfig.getNgEnvironmentGlobalOverride();
-    ServiceOverridesSpecBuilder specBuilder =
-        ServiceOverridesSpec.builder().variables(ngEnvironmentInfoConfig.getVariables());
-    if (ngEnvironmentGlobalOverride != null) {
-      specBuilder.manifests(ngEnvironmentGlobalOverride.getManifests())
-          .configFiles(ngEnvironmentGlobalOverride.getConfigFiles())
-          .connectionStrings(ngEnvironmentGlobalOverride.getConnectionStrings())
-          .applicationSettings(ngEnvironmentGlobalOverride.getApplicationSettings());
-    }
-
-    final String envRef =
-        IdentifierRefHelper.getRefFromIdentifierOrRef(accountId, ngEnvironmentInfoConfig.getOrgIdentifier(),
-            ngEnvironmentInfoConfig.getProjectIdentifier(), ngEnvironmentInfoConfig.getIdentifier());
-
-    return NGServiceOverrideConfigV2.builder()
-        .identifier(generateEnvGlobalOverrideV2Identifier(envRef))
-        .environmentRef(envRef)
-        .type(ENV_GLOBAL_OVERRIDE)
-        .spec(specBuilder.build())
-        .build();
-  }
-
-  private String generateEnvServiceOverrideV2Identifier(String envRef, String serviceRef) {
-    return String.join("_", envRef, serviceRef).replace(".", "_");
-  }
-
-  private String generateEnvGlobalOverrideV2Identifier(String envRef) {
-    return String.join("_", envRef).replace(".", "_");
-  }
-
-  @Data
-  @Builder
-  static class ServicePartResponse {
-    NGServiceConfig ngServiceConfig;
   }
 }
